requirements.txt
fastapi
uvicorn[standard]
celery
redis
psycopg2-binary
opencv-python-headless
numpy
scikit-learn
insightface
faiss-cpu
python-multipart
pillow
onnxruntime
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
razorpay
pydantic[email]